@@ -22,15 +22,15 @@
         <developerConnection>scm:git:ssh://git@github.com:entur/kakka.git</developerConnection>
     </scm>
     <properties>
-        <java.version>17</java.version>
-        <camel.version>3.20.0</camel.version>
+        <java.version>21</java.version>
+        <camel.version>4.3.0</camel.version>
 
         <wololo.version>0.14.3</wololo.version>
         <jskdata.version>1.0.15</jskdata.version>
 
         <auth0.version>1.26.0</auth0.version>
-        <netex-java-model.version>1.0.12-SNAPSHOT</netex-java-model.version>
-        <entur.helpers.version>1.97</entur.helpers.version>
+        <netex-java-model.version>2.0.14</netex-java-model.version>
+        <entur.helpers.version>2.7</entur.helpers.version>
         <guava.version>31.1-jre</guava.version>
 
         <entur.google.pubsub.emulator.download.skip>false</entur.google.pubsub.emulator.download.skip>
