@@ -1,11 +1,12 @@
 package no.entur.kakka.security;
 
-import org.entur.oauth2.RorAuthenticationConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.stereotype.Component;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
@@ -17,7 +18,7 @@ import static org.springframework.security.config.Customizer.withDefaults;
 
 /**
  * Authentication and authorization configuration for Kakka.
- * All requests must be authenticated except for the Swagger and Actuator endpoints.
+ * All requests must be authenticated except for the OpenAPI and Actuator endpoints.
  */
 @Profile("!test")
 @EnableWebSecurity
@@ -36,26 +37,23 @@ public class KakkaWebSecurityConfigurerAdapter{
     }
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+    public SecurityFilterChain filterChain(HttpSecurity http, MultiIssuerAuthenticationManagerResolver multiIssuerAuthenticationManagerResolver) throws Exception {
         http.cors(withDefaults())
-                .csrf().disable()
-                .authorizeRequests()
-                .antMatchers("/services/swagger.json").permitAll()
-                .antMatchers("/services/geocoder_admin/swagger.json").permitAll()
-                .antMatchers("/services/organisation_admin/swagger.json").permitAll()
-                .antMatchers("/services/export/swagger.json").permitAll()
-                .antMatchers("/services/osmpoifilter").permitAll()
-                // exposed internally only, on a different port (pod-level)
-                .antMatchers("/actuator/prometheus").permitAll()
-                .antMatchers("/actuator/health").permitAll()
-                .antMatchers("/actuator/health/liveness").permitAll()
-                .antMatchers("/actuator/health/readiness").permitAll()
-                .anyRequest().authenticated()
-                .and()
-                .oauth2ResourceServer().jwt().jwtAuthenticationConverter(new RorAuthenticationConverter())
-                .and()
-                .and()
-                .oauth2Client();
+                .csrf(AbstractHttpConfigurer::disable)
+                .authorizeRequests(authz -> authz
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/services/openapi.json")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/services/geocoder_admin/openapi.json")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/services/organisation_admin/openapi.json")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/services/export/openapi.json")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/services/osmpoifilter")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/actuator/prometheus")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/actuator/health")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/actuator/health/liveness")).permitAll()
+                        .requestMatchers(AntPathRequestMatcher.antMatcher("/actuator/health/readiness")).permitAll()
+                        .anyRequest().authenticated())
+                .oauth2ResourceServer(configurer -> configurer.authenticationManagerResolver(multiIssuerAuthenticationManagerResolver))
+                .oauth2Client(withDefaults());
+
 
         return http.build();
 
