@@ -2,7 +2,7 @@ package no.entur.kakka.security;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
-import org.apache.http.HttpHeaders;
+import org.apache.hc.core5.http.HttpHeaders;
 import org.entur.oauth2.TokenService;
 import org.springframework.stereotype.Component;
 
