@@ -10,6 +10,7 @@ import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 
 
@@ -31,7 +32,7 @@ public class AdminRestRouteBuilderTest extends KakkaRouteBuilderIntegrationTestB
 
         @Bean
         public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-            http.csrf().disable()
+            http.csrf(AbstractHttpConfigurer::disable)
                     .authorizeRequests(authorizeRequests ->
                             authorizeRequests
                                     .anyRequest().permitAll()
